+++ sdk/monitor/azure-monitor-opentelemetry-exporter/CHANGELOG.md
@@ -15,6 +15,9 @@
 
 ### Other Changes
 
+- Refactored live metrics filtering modules
+  ([#37694](https://github.com/Azure/azure-sdk-for-python/pull/37694))
+
 ## 1.0.0b32 (2024-11-04)
 
 ### Breaking Changes